requirements.txt
flask>=3.0.0
nltk>=3.8.0
Pillow>=10.0.0
torch>=2.1.0
diffusers>=0.26.0
transformers>=4.38.0
accelerate>=0.27.0
